Add clustering parameters to scanpy QC.
