+++ api/tools/qc/scanpy_qc.py
@@ -7,12 +7,13 @@ warnings.filterwarnings('ignore')
 from scipy.stats import median_abs_deviation
 from tools.formating.formating import is_normalized, check_nonnegative_integers
 from scipy.sparse import csr_matrix
+from tools.utils.reduction import run_dimension_reduction
 sc.settings.verbosity=3             # verbosity: errors (0), warnings (1), info (2), hints (3)
 sc.logging.print_header()
 # sc.settings.set_figure_params(dpi=80, facecolor='white')
 
 
-def run_scanpy_qc(adata, min_genes=200, max_genes=None, min_cells=3, target_sum=1e4, n_top_genes=None ,regress_cell_cycle=False):
+def run_scanpy_qc(adata, min_genes=200, max_genes=None, min_cells=3, target_sum=1e4, n_top_genes=None, n_neighbors=10, n_pcs=40, resolution=1, regress_cell_cycle=False):
         if adata is None:
             raise ValueError("The input is None.")
         
@@ -92,6 +93,8 @@ def run_scanpy_qc(adata, min_genes=200, max_genes=None, min_cells=3, target_sum=
             adata.X = csr_matrix(adata.X)
 
         adata.layers["log10k"] = adata.X.copy()
+
+        adata = run_dimension_reduction(adata, n_neighbors=n_neighbors, n_pcs=n_pcs, resolution=resolution)
 
         # return adata, output
         return adata
